+++ src/javascript/main.js
@@ -30,6 +30,7 @@ require('../modules/block-content-newsletter/script');
 require('../modules/button-top/script');
 require('../modules/gallery-ugc/script');
 require('../modules/icon-switch/script');
+require('../modules/jumbotron-slider/script');
 require('../modules/notice/script');
 require('../modules/person-block/script');
 require('../modules/poll/script');
@@ -52,5 +53,5 @@ require('./lib/egg');
 require('./lib/3975');
 
 var Paris = window.Paris || {};
-Paris.version = "1.7.5";
+Paris.version = "1.8.0";
 Paris.templates = require('./client.tpl');